backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-jose[cryptography]
bcrypt==4.0.1
passlib==1.7.4
httpx
pydantic
email-validator
pydantic-settings
reportlab